requirements.txt
alembic==1.13.3
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
DateTime==5.5
distlib==0.3.8
dnspython==2.6.1
email_validator==2.2.0
filelock==3.16.1
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-Redmail @ git+https://github.com/shetty-777/flask-redmail-g@32f7c9ec5c072fedf29f87fb0795e7b0ca0c6c56
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
pip-review==1.3.0
platformdirs==4.3.6
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.2
redmail==0.6.0
setuptools==75.1.0
soupsieve==2.6
SQLAlchemy==2.0.35
typing_extensions==4.12.2
tzdata==2024.2
virtualenv==20.26.6
waitress==3.0.0
Werkzeug==3.0.4
wheel==0.44.0
WTForms==3.1.2
zope.interface==7.0.3
